chore(auth): Backport Gen2 config to Gen1 JSON for older authenticator versions (#2768)
