+++ aws-auth-cognito/src/main/java/com/amplifyframework/statemachine/codegen/data/OauthConfiguration.kt
@@ -16,6 +16,7 @@
 package com.amplifyframework.statemachine.codegen.data
 
 import com.amplifyframework.annotations.InternalAmplifyApi
+import org.json.JSONArray
 import org.json.JSONObject
 
 @InternalAmplifyApi
@@ -27,22 +28,38 @@ data class OauthConfiguration internal constructor(
     val signInRedirectURI: String,
     val signOutRedirectURI: String
 ) {
+    internal fun toGen1Json() = JSONObject().apply {
+        put(AppClientId, appClient)
+        appSecret?.let { put(AppClientSecret, it) }
+        put(WebDomain, domain)
+        put(Scopes, JSONArray(scopes))
+        put(SignInRedirectURI, signInRedirectURI)
+        put(SignOutRedirectURI, signOutRedirectURI)
+    }
+
     internal companion object {
+
+        private const val AppClientId = "AppClientId"
+        private const val AppClientSecret = "AppClientSecret"
+        private const val WebDomain = "WebDomain"
+        private const val Scopes = "Scopes"
+        private const val SignInRedirectURI = "SignInRedirectURI"
+        private const val SignOutRedirectURI = "SignOutRedirectURI"
 
         fun fromJson(jsonObject: JSONObject?): OauthConfiguration? {
             return jsonObject?.run {
-                val appClient = optString("AppClientId").takeUnless { it.isNullOrEmpty() }
-                val appSecret = optString("AppClientSecret", null).takeUnless { it.isNullOrEmpty() }
-                val domain = optString("WebDomain").takeUnless { it.isNullOrEmpty() }
-                val scopes = optJSONArray("Scopes")?.let { scopesArray ->
+                val appClient = optString(AppClientId).takeUnless { it.isNullOrEmpty() }
+                val appSecret = optString(AppClientSecret, null).takeUnless { it.isNullOrEmpty() }
+                val domain = optString(WebDomain).takeUnless { it.isNullOrEmpty() }
+                val scopes = optJSONArray(Scopes)?.let { scopesArray ->
                     val scopesSet = mutableSetOf<String>()
-                    for (i in 0..scopesArray.length()) {
+                    for (i in 0 until scopesArray.length()) {
                         scopesArray.optString(i)?.let { scopesSet.add(it) }
                     }
                     scopesSet
                 }
-                val signInRedirectURI = optString("SignInRedirectURI").takeUnless { it.isNullOrEmpty() }
-                val signOutRedirectURI = optString("SignOutRedirectURI").takeUnless { it.isNullOrEmpty() }
+                val signInRedirectURI = optString(SignInRedirectURI).takeUnless { it.isNullOrEmpty() }
+                val signOutRedirectURI = optString(SignOutRedirectURI).takeUnless { it.isNullOrEmpty() }
 
                 return if (appClient != null && domain != null && scopes != null && signInRedirectURI != null &&
                     signOutRedirectURI != null
